IGeometryGrid API Update, Pythong Fixes, MXA Removal
The API is changed to use a FloatVec3Type or SizeVec3Type (both alias for SIMPLArray<T, N>) so that
the API is consistent among the various classes and easy to use within SIMPL/DREAM.3D and its plugins.
The use of the std::tuple<T> based return types is deprecated and removed.

The Python script files were completely reorganized to put the scripts with their plugins. Most of
the scripts were tested on a macOS machine and added to the ctest suite of unit tests. For each script
a bash script is created that drives the test. The Python tests are dependant on an Anaconda3 installation
being used. A few of the Python tests are still failing but those will be fixed eventually

The source code was searched for the old and deprecated MXA strings and classes and the old code was
either removed or renamed to SIMPL.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>